refactor(provider): consolidate field resource logic and optimize CRUD hooks handling
- Centralized `CRUDRunner` utilization for `fieldResource` by introducing a shared instance.
- Simplified field state mapping and API payload construction via new helper functions.
- Removed redundant `fieldResourceModel` implementation from outdated `field_types.go`.
- Enhanced error handling and state mapping for `jira_field` resource.
